Weekly Devlog #6 & 7
Finally got grass working so that you actually get decent frames when rendering it all, used a geometry shader and its done wonders. Also added wind to the grass and grass is pretty much done now. Added object spawning so that I could spawn trees, also stonehenge has a 1/400 chance to spawn on any given chunk. Allows me to customize the frequency of that structure spawn using chunks and allows me to tell it how many of that object are on a chunk that does contain whatever it is. Right now trees have a 50% chance of spawning on any given chunk and if they do, 10 of them spawn. Makes the trees look really natural but I also plan on improving that system later on. Next is making chunks generate around the player and adding save data for chunks. By week 9 or 10 I hope to be at biome/water implementation but only time will tell.