Weekly Devlog #9
Finally able to get back to work after exams and competitions, changed the character controller so that instead of moving the player it moves the terrain as to prevent floating point error problems furthur down the line with infinitely generated terrain. Not a week of development more like a day, going to my moms tomorrow and going to start on the infinite generation around the player part of the game.